Medieval Overhaul Fix
- Removed a number of patches for Medieval Overhaul to add duplicate tasks to Production.  The patches are for Millstone, Spinning Wheel and Trestle.

Note: The Production worktype is in a bit of an odd place.  Ideally it should be for all production tasks but the problem with that is certain things are attached to actual skills.  This can create the issue where a pawn can do something they should be incapable of doing making them work super slowly and try to use a work speed stat they don't have causing the game to throw an error.  Removing the skill stuff means a lot of patching and losing the XP gain these things provide.  It's an annoying and difficult problem which doesn't have a perfect solution.  With that in mind I'm thinking that it would probably be best to shift Production toward "Skillless Dumb Labour".  It will be less useful but with mods should still be useful enough to split off as a worktype.  It would also resolve the myriad of issues the current setup causes.